move types and tests into src
