+++ clients/validator/package.json
@@ -6,11 +6,11 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "build": "rollup -c rollup.config.mjs",
+    "build": "rollup -c ./rollup.config.mjs",
     "build:types": "rollup-type-bundler --dist ./dist",
     "build:prod": "sh ./scripts/build-prod.sh",
-    "test": "ts-mocha -p ./tests/tsconfig.test.json tests/**/*.test.ts",
-    "testmock": "ts-mocha -p ./tests/tsconfig.test.json tests/mock/*.test.ts",
+    "test": "ts-mocha -p ./tsconfig.test.json ./src/tests/**/*.test.ts",
+    "testmock": "ts-mocha -p ./tsconfig.test.json ./src/tests/mock/*.test.ts",
     "coverage": "nyc npm test",
     "clean": "rm -rf ./dist",
     "lint": "eslint",
