+++ clients/validator/rollup.config.mjs
@@ -2,20 +2,14 @@ import typescript from '@rollup/plugin-typescript';
 import resolve from '@rollup/plugin-node-resolve';
 import json from '@rollup/plugin-json';
 import commonjs from '@rollup/plugin-commonjs';
-import dts from 'rollup-plugin-dts';
 
 export default [
   {
-    input: 'src/index.ts',
+    input: './src/index.ts',
     output: {
       dir: 'dist',
       format: 'cjs',
     },
     plugins: [resolve(), typescript(), commonjs(), json()],
-  },
-  {
-    input: './dist/index.d.ts',
-    output: [{ file: 'dist/types.d.ts' }],
-    plugins: [dts()],
   },
 ];
