+++ clients/validator/src/tests/sign/mixnetActions.test.ts
@@ -1,5 +1,5 @@
 import expect from 'expect';
-import ValidatorClient from '../../src';
+import ValidatorClient from '../../';
 
 const dotenv = require('dotenv');
