+++ clients/validator/src/types/global.d.ts
